requirements.txt
python-whois
tldextract
google-search-results